Initial algorithm

+++ SchoolBusRouting/SchoolBusRouting/Algorithm/GeneticAlgorithm.cs
@@ -11,12 +11,12 @@ namespace SchoolBusRouting.Algorithm
         public ICrossover<T> Crossover { get; }
         public IFitnessFunction<T> FitnessFunction { get; }
         public int IterationLimit { get; }
-        public decimal FitnessTerminator { get; }
+        public double FitnessTerminator { get; }
         public int PopulationSize { get; }
         protected Population<T> Population { get; set; }
 
         protected GeneticAlgorithm(IMutation<T> mutation, ISelection<T> selection, ICrossover<T> crossover, 
-            IFitnessFunction<T> fitnessFunction, int iterationLimit, decimal fitnessTerminator, int populationSize)
+            IFitnessFunction<T> fitnessFunction, int iterationLimit, double fitnessTerminator, int populationSize)
         {
             Mutation = mutation;
             Selection = selection;
